# Application Programming Interface/BuildingBug.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BuildingBugAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/v{version}/Servico/new": {
      "post": {
        "tags": [
          "Servico"
        ],
        "summary": "Pedir serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/update": {
      "put": {
        "tags": [
          "Servico"
        ],
        "summary": "Atualizar serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/delete/{id}": {
      "delete": {
        "tags": [
          "Servico"
        ],
        "summary": "Eliminar serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/{id}": {
      "get": {
        "tags": [
          "Servico"
        ],
        "summary": "Obter serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/all": {
      "get": {
        "tags": [
          "Servico"
        ],
        "summary": "Obter todos os serviços",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/all/{profissionalId}": {
      "get": {
        "tags": [
          "Servico"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profissionalId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/all/Cliente/{clienteId}": {
      "get": {
        "tags": [
          "Servico"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/all/disponivel": {
      "get": {
        "tags": [
          "Servico"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/aceitar/{profissionalId}/{id}": {
      "put": {
        "tags": [
          "Servico"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profissionalId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/new": {
      "post": {
        "tags": [
          "Tipo"
        ],
        "summary": "Criar Tipo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTipoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/update": {
      "put": {
        "tags": [
          "Tipo"
        ],
        "summary": "Atualizar Tipo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTipoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/delete/{id}": {
      "delete": {
        "tags": [
          "Tipo"
        ],
        "summary": "Eliminar Tipo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/{id}": {
      "get": {
        "tags": [
          "Tipo"
        ],
        "summary": "Obter Tipo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/all": {
      "get": {
        "tags": [
          "Tipo"
        ],
        "summary": "Obter todos os Tipos",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Tipo/categoria/{categoria}": {
      "get": {
        "tags": [
          "Tipo"
        ],
        "summary": "Obter todos os tipos de uma determinada categoria",
        "parameters": [
          {
            "in": "path",
            "name": "categoria",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateServicoRequestContract": {
      "required": [
        "clienteId",
        "codigoPostalCP",
        "fator",
        "morada",
        "tipoId"
      ],
      "type": "object",
      "properties": {
        "morada": {
          "minLength": 1,
          "type": "string"
        },
        "fator": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "codigoPostalCP": {
          "format": "int32",
          "type": "integer"
        },
        "tipoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "CreateTipoRequestContract": {
      "type": "object",
      "properties": {
        "descricao": {
          "type": "string"
        },
        "preco": {
          "format": "int32",
          "type": "integer"
        },
        "fator": {
          "type": "boolean"
        },
        "categoria": {
          "type": "string"
        }
      }
    },
...
    "UpdateServicoRequestContract": {
      "required": [
        "avaliacao",
        "clienteId",
        "codigoPostalCP",
        "custo",
        "dataFim",
        "dataInicio",
        "estadoServicoId",
        "fator",
        "id",
        "morada",
        "profissionalId",
        "tipoId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "morada": {
          "minLength": 1,
          "type": "string"
        },
        "dataInicio": {
          "format": "date-time",
          "type": "string"
        },
        "dataFim": {
          "format": "date-time",
          "type": "string"
        },
        "avaliacao": {
          "format": "int32",
          "type": "integer"
        },
        "custo": {
          "format": "float",
          "type": "number"
        },
        "fator": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "estadoServicoId": {
          "format": "int32",
          "type": "integer"
        },
        "codigoPostalCP": {
          "format": "int32",
          "type": "integer"
        },
        "profissionalId": {
          "format": "int32",
          "type": "integer"
        },
        "tipoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateTipoRequestContract": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descricao": {
          "type": "string"
        },
        "preco": {
          "format": "int32",
          "type": "integer"
        },
        "fator": {
          "type": "boolean"
        },
        "categoria": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insert Token as: 'Bearer your_token_here'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}